Virtual hard drive init
This script adds a virtual hard disks to the current path. It deletes the existing one if one is found at the same location.

TODO: Ask user to delete VDI if one is found
TODO: Mechanism for input like filename and diks size
TODO: Feedback in the form of UUID for VDI so that can be used instead of name, which may be duplicate